fix: Si no existe el archivo de sesión no borrar al cerrar la aplicación

+++ src/functions.py
@@ -5,5 +5,6 @@ def exit_application(msg):
     load_dotenv()
     print(msg)
     # Remove session file
-    os.remove(os.getenv("SESSION_FILE"))
+    if os.path.exists(os.getenv("SESSION_FILE")):
+        os.remove(os.getenv("SESSION_FILE"))
     exit()
